requirements.txt
django
django-phonenumber-field
django-phonenumbers
django-select2
django-jazzmin
django-money
django-measurement
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
pandas
requests
openpyxl-image-loader
openpyxl
